move symbol view related code to c2c-symbol.c
