+++ tools/perf/builtin-c2c.c
@@ -53,120 +53,11 @@
 #include "string2.h"
 #include "util/util.h"
 
-static inline bool symbol_name_equal(struct symbol *a, struct symbol *b)
-{
-	return a && b && strcmp(a->name, b->name) == 0;
-}
-
-struct c2c_hists {
-	struct hists		hists;
-	struct perf_hpp_list	list;
-	struct c2c_stats	stats;
-};
-
-struct compute_stats {
-	struct stats		 lcl_hitm;
-	struct stats		 rmt_hitm;
-	struct stats		 lcl_peer;
-	struct stats		 rmt_peer;
-	struct stats		 load;
-};
-
-struct c2c_hist_entry {
-	struct c2c_hists	*hists;
-	struct c2c_stats	 stats;
-	unsigned long		*cpuset;
-	unsigned long		*nodeset;
-	struct c2c_stats	*node_stats;
-	unsigned int		 cacheline_idx;
-
-	struct compute_stats	 cstats;
-
-	unsigned long		 paddr;
-	unsigned long		 paddr_cnt;
-	bool			 paddr_zero;
-	char			*nodestr;
-
-	/* Symbol association support */
-	struct list_head	related_symbols;	/* Related symbols list */
-
-	/* Cached total cycles for this entry to avoid repeated calculations */
-	uint64_t		 total_cycles;
-	bool			 total_cycles_valid;
-
-	/*
-	 * must be at the end,
-	 * because of its callchain dynamic entry
-	 */
-	struct hist_entry	he;
-};
+/*
+ * Structures are now defined in builtin-c2c.h
+ */
 
 static char const *coalesce_default = "iaddr";
-
-/* Structure to represent related symbols */
-struct related_symbol {
-	struct list_head	 list;
-	struct symbol		*sym;
-	uint64_t		iaddr;			/* Code address */
-	struct c2c_stats	stats;			/* Aggregated stats */
-	struct compute_stats	cstats;			/* Compute stats */
-};
-
-/* Cacheline-to-symbols index entry for performance optimization */
-struct cacheline_symbol_entry {
-	struct hist_entry *he_cl;
-	struct c2c_hist_entry *c2c_he_cl;
-	struct symbol_access {
-		struct symbol *sym;
-		uint64_t iaddr;
-		struct map *map;
-		struct maps *maps;
-		struct c2c_stats stats;
-		struct compute_stats cstats;
-		struct symbol_access *next;
-	} *symbol_accesses;
-};
-
-struct perf_c2c {
-	struct perf_tool	tool;
-	struct c2c_hists	hists;
-	struct c2c_hists	symbol_hists;  /* Symbol grouped hists */
-	struct mem2node		mem2node;
-
-	unsigned long		**nodes;
-	int			 nodes_cnt;
-	int			 cpus_cnt;
-	int			*cpu2node;
-	int			 node_info;
-
-	bool			 show_src;
-	bool			 show_all;
-	bool			 use_stdio;
-	bool			 stats_only;
-	bool			 symbol_full;
-	bool			 stitch_lbr;
-
-	/* Shared cache line stats */
-	struct c2c_stats	shared_clines_stats;
-	int			shared_clines;
-
-	int			 display;
-
-	/* Cached total cycles across all symbols for percent column */
-	uint64_t		symbol_total_cycles;
-	bool			symbol_total_cycles_valid;
-
-	const char		*coalesce;
-	char			*cl_sort;
-	char			*cl_resort;
-	char			*cl_output;
-
-	/* Cacheline-to-symbols index for performance optimization */
-	struct cacheline_symbol_entry *cacheline_index;
-	int			cacheline_index_size;
-	int			cacheline_index_capacity;
-	bool			cacheline_index_built;
-};
 
 enum {
 	DISPLAY_LCL_HITM,
@@ -190,19 +81,10 @@ static const struct option c2c_options[] = {
 
 struct perf_c2c c2c;
 
-static int build_symbol_hists(struct perf_env *env);
-
-/* Helper function to initialize c2c_hist_entry related_symbols */
-static void init_c2c_he_related_symbols(struct c2c_hist_entry *c2c_he)
-{
-	INIT_LIST_HEAD(&c2c_he->related_symbols);
-}
-
-/* Helper function to invalidate cached total cycles */
-static void c2c_he_invalidate_total_cycles_cache(struct c2c_hist_entry *c2c_he)
-{
-	c2c_he->total_cycles_valid = false;
-}
+/*
+ * Helper functions init_c2c_he_related_symbols and c2c_he_invalidate_total_cycles_cache
+ * are now inline functions in builtin-c2c.h
+ */
 
 static void *c2c_he_zalloc(size_t size)
 {
@@ -331,15 +213,13 @@ static void c2c_he_free(void *he)
 	free(c2c_he);
 }
 
-static struct hist_entry_ops c2c_entry_ops = {
+/* Exported for c2c-symbol.c */
+struct hist_entry_ops c2c_entry_ops = {
 	.new	= c2c_he_zalloc,
 	.free	= c2c_he_free,
 };
 
-static int c2c_hists__init(struct c2c_hists *hists,
-			   const char *sort,
-			   int nr_header_lines,
-			   struct perf_env *env);
+/* c2c_hists__init is declared in builtin-c2c.h */
 
 static struct c2c_hists*
 he__get_c2c_hists(struct hist_entry *he,
@@ -443,8 +323,8 @@ static void merge_stats(struct stats *dest, struct stats *src)
 		dest->min = src->min;
 }
 
-/* Function to merge compute_stats during symbol aggregation */
-static void c2c_add_cstats(struct compute_stats *dest, struct compute_stats *src)
+/* Function to merge compute_stats during symbol aggregation - exported for c2c-symbol.c */
+void c2c_add_cstats(struct compute_stats *dest, struct compute_stats *src)
 {
 	merge_stats(&dest->rmt_hitm, &src->rmt_hitm);
 	merge_stats(&dest->lcl_hitm, &src->lcl_hitm);
@@ -2745,10 +2625,11 @@ static int hpp_list__parse(struct perf_hpp_list *hpp_list,
 	return ret;
 }
 
-static int c2c_hists__init(struct c2c_hists *hists,
-			   const char *sort,
-			   int nr_header_lines,
-			   struct perf_env *env)
+/* Exported for c2c-symbol.c */
+int c2c_hists__init(struct c2c_hists *hists,
+		    const char *sort,
+		    int nr_header_lines,
+		    struct perf_env *env)
 {
 	__hists__init(&hists->hists, &hists->list);
 
@@ -2765,10 +2646,11 @@ static int c2c_hists__init(struct c2c_hists *hists,
 	return hpp_list__parse(&hists->list, /*output=*/NULL, sort, env);
 }
 
-static int c2c_hists__reinit(struct c2c_hists *c2c_hists,
-			     const char *output,
-			     const char *sort,
-			     struct perf_env *env)
+/* Exported for c2c-symbol.c */
+int c2c_hists__reinit(struct c2c_hists *c2c_hists,
+		      const char *output,
+		      const char *sort,
+		      struct perf_env *env)
 {
 	perf_hpp__reset_output_field(&c2c_hists->list);
 	return hpp_list__parse(&c2c_hists->list, output, sort, env);
@@ -3263,889 +3145,16 @@ static void perf_c2c__hists_fprintf(FILE *out, struct perf_session *session)
 
 #ifdef HAVE_SLANG_SUPPORT
 
-/* Comparison function for sorting related symbols by stats.store descending */
-static int related_symbol_cmp(const void *a, const void *b)
-{
-	const struct related_symbol *sym_a = *(const struct related_symbol **)a;
-	const struct related_symbol *sym_b = *(const struct related_symbol **)b;
-
-	if (sym_b->stats.store > sym_a->stats.store)
-		return 1;
-	else if (sym_b->stats.store < sym_a->stats.store)
-		return -1;
-	return 0;
-}
-
-struct grand_item {
-	struct c2c_hist_entry *grand_c2c;
-	struct hist_entry *grand_he;
-	u64 stores;
-};
-
-static int grand_item_cmp(const void *a, const void *b)
-{
-	const struct grand_item *item_a = (const struct grand_item *)a;
-	const struct grand_item *item_b = (const struct grand_item *)b;
-
-	if (item_b->stores > item_a->stores)
-		return 1;
-	else if (item_b->stores < item_a->stores)
-		return -1;
-	return 0;
-}
-
 /*
- * Helper function to validate and prepare for populating symbol children
- * Returns the c2c_hist_entry if valid, NULL otherwise
+ * Symbol view processing functions are now implemented in
+ * ui/browsers/c2c-symbol.c:
+ *   - build_symbol_hists()
+ *   - build_cacheline_symbol_index()
+ *   - cleanup_cacheline_symbol_index()
+ *   - populate_symbol_children()
+ *   - build_symbol_associations()
  */
-static struct c2c_hist_entry *validate_and_prepare_entries(struct hist_entry *he)
-{
-	struct c2c_hist_entry *c2c_he;
-	struct rb_root_cached *root;
 
-	if (!he || !he->has_children)
-		return NULL;
-
-	root = &he->hroot_out;
-
-	/* If already populated, return */
-	if (!RB_EMPTY_ROOT(&root->rb_root))
-		return NULL;
-
-	c2c_he = container_of(he, struct c2c_hist_entry, he);
-	if (!c2c_he)
-		return NULL;
-
-	/* Ensure related_symbols list is valid */
-	if (list_empty(&c2c_he->related_symbols)) {
-		he->has_children = false;  /* Reset inconsistent state */
-		return NULL;
-	}
-
-	return c2c_he;
-}
-
-/*
- * Helper function to sort related symbols by stores descending
- * Returns allocated array of sorted related_symbol pointers, or NULL on error
- * Caller must free the returned array
- */
-static struct related_symbol **sort_related_symbols_by_stores(struct c2c_hist_entry *c2c_he, int *num_rel_out)
-{
-	struct related_symbol *rel_sym;
-	struct related_symbol **sorted = NULL;
-	int num_rel = 0, idx = 0;
-
-	/* Count related symbols and allocate array in one pass */
-	list_for_each_entry(rel_sym, &c2c_he->related_symbols, list)
-		num_rel++;
-
-	if (num_rel == 0) {
-		*num_rel_out = 0;
-		return NULL;
-	}
-
-	sorted = calloc(num_rel, sizeof(*sorted));
-	if (!sorted) {
-		*num_rel_out = 0;
-		return NULL;
-	}
-
-	/* Fill array in the same pass */
-	list_for_each_entry(rel_sym, &c2c_he->related_symbols, list)
-		sorted[idx++] = rel_sym;
-
-	/* Sort by stats.store descending */
-	qsort(sorted, num_rel, sizeof(*sorted), related_symbol_cmp);
-
-	*num_rel_out = num_rel;
-	return sorted;
-}
-
-/*
- * Helper function to create and initialize a symbol child entry
- * Returns the created hist_entry or NULL on error
- */
-static struct hist_entry *create_symbol_child_entry(struct hist_entry *parent_he,
-						      struct related_symbol *rel_sym)
-{
-	struct c2c_hist_entry *child_c2c_he, *child_c2c;
-	struct hist_entry *child_he;
-
-	if (!rel_sym || !rel_sym->sym)
-		return NULL;
-
-	/* Allocate child hist_entry - simplified version for symbol children */
-	child_c2c_he = zalloc(sizeof(*child_c2c_he));
-	if (!child_c2c_he)
-		return NULL;
-
-	/* Initialize the related_symbols list */
-	init_c2c_he_related_symbols(child_c2c_he);
-
-	child_he = &child_c2c_he->he;
-
-	/* Complete initialization - copy parent's map_symbol structure first */
-	memcpy(&child_he->ms, &parent_he->ms, sizeof(struct map_symbol));
-	/* Then override the symbol and address */
-	child_he->ms.sym = rel_sym->sym;
-
-	/* Create a synthetic mem_info to store the iaddr for proper display */
-	if (parent_he->mem_info) {
-		child_he->mem_info = mem_info__clone(parent_he->mem_info);
-		if (child_he->mem_info) {
-			/* Set the instruction address to the related symbol's iaddr */
-			mem_info__iaddr(child_he->mem_info)->addr = rel_sym->iaddr;
-		}
-	}
-
-	/* Copy essential fields from parent */
-	child_he->thread = parent_he->thread;
-	child_he->cpumode = parent_he->cpumode;
-	child_he->cpu = parent_he->cpu;
-	child_he->socket = parent_he->socket;
-
-	/* Set hierarchy fields */
-	child_he->parent_he = parent_he;
-	child_he->depth = parent_he->depth + 1;
-	child_he->leaf = false; /* it can have grandchildren (cachelines) */
-	/* In symbol view, all entries should use symbol_hists */
-	child_he->hists = &c2c.symbol_hists.hists;
-	child_he->filtered = false;  /* Make sure it's not filtered out */
-	child_he->unfolded = false;
-	child_he->has_children = false; /* Will be set to true only if grandchildren are added */
-	child_he->has_no_entry = false;
-	child_he->nr_rows = 0;
-	child_he->row_offset = 0;
-
-	/* Initialize stats properly */
-	memset(&child_he->stat, 0, sizeof(child_he->stat));
-
-	/* Set stat values based on c2c stats */
-	child_he->stat.nr_events = rel_sym->stats.rmt_hitm + rel_sym->stats.lcl_hitm +
-				   rel_sym->stats.rmt_peer + rel_sym->stats.lcl_peer;
-	child_he->stat.period = child_he->stat.nr_events;
-
-	/* These weight fields are used by some columns */
-	child_he->stat.weight1 = rel_sym->stats.rmt_hitm + rel_sym->stats.lcl_hitm;
-
-	/* Initialize stat_acc - allocate if needed */
-	if (symbol_conf.cumulate_callchain) {
-		child_he->stat_acc = calloc(1, sizeof(struct he_stat));
-		if (child_he->stat_acc)
-			memcpy(child_he->stat_acc, &child_he->stat, sizeof(struct he_stat));
-	}
-
-	/* Initialize rb trees */
-	child_he->hroot_in = RB_ROOT_CACHED;
-	child_he->hroot_out = RB_ROOT_CACHED;
-
-	/* Initialize pairs list */
-	INIT_LIST_HEAD(&child_he->pairs.node);
-
-	/* Initialize hierarchy pointers */
-	child_he->hpp_list = &c2c.symbol_hists.list;
-
-	/* Copy c2c stats - this is what c2c columns use */
-	child_c2c = container_of(child_he, struct c2c_hist_entry, he);
-	memcpy(&child_c2c->stats, &rel_sym->stats, sizeof(rel_sym->stats));
-	memcpy(&child_c2c->cstats, &rel_sym->cstats, sizeof(rel_sym->cstats));
-	init_c2c_he_related_symbols(child_c2c);
-
-	/* Build cacheline grandchildren under each related symbol child */
-	child_c2c->hists = NULL;
-
-	return child_he;
-}
-
-/* Build index mapping cachelines to accessing symbols for performance */
-void build_cacheline_symbol_index(void)
-{
-	struct rb_node *nd_cl;
-	int index = 0;
-
-	/* Return early if already built */
-	if (c2c.cacheline_index_built)
-		return;
-
-	/* Free existing index if any */
-	if (c2c.cacheline_index) {
-		for (int i = 0; i < c2c.cacheline_index_size; i++) {
-			struct symbol_access *sa = c2c.cacheline_index[i].symbol_accesses;
-
-			while (sa) {
-				struct symbol_access *next = sa->next;
-
-				free(sa);
-				sa = next;
-			}
-		}
-		free(c2c.cacheline_index);
-		c2c.cacheline_index = NULL;
-	}
-
-	/* Build index in single pass with dynamic array growth */
-	c2c.cacheline_index_size = 0;
-	c2c.cacheline_index_capacity = 64; /* Start with reasonable size */
-
-	c2c.cacheline_index = malloc(c2c.cacheline_index_capacity * sizeof(struct cacheline_symbol_entry));
-	if (!c2c.cacheline_index) {
-		c2c.cacheline_index_size = 0;
-		return;
-	}
-
-	nd_cl = rb_first_cached(&c2c.hists.hists.entries);
-	while (nd_cl) {
-		struct hist_entry *he_cl;
-		struct c2c_hist_entry *c2c_he_cl;
-
-		/* Grow array if needed */
-		if (index >= c2c.cacheline_index_capacity) {
-			struct cacheline_symbol_entry *new_index;
-			int cleanup_i;
-
-			c2c.cacheline_index_capacity *= 2;
-			new_index = realloc(c2c.cacheline_index,
-				c2c.cacheline_index_capacity * sizeof(struct cacheline_symbol_entry));
-			if (!new_index) {
-				/* Cleanup on allocation failure */
-				for (cleanup_i = 0; cleanup_i < c2c.cacheline_index_size; cleanup_i++) {
-					struct symbol_access *sa = c2c.cacheline_index[cleanup_i].symbol_accesses;
-
-					while (sa) {
-						struct symbol_access *next = sa->next;
-
-						free(sa);
-						sa = next;
-					}
-				}
-				free(c2c.cacheline_index);
-				c2c.cacheline_index = NULL;
-				c2c.cacheline_index_size = 0;
-				return;
-			}
-			c2c.cacheline_index = new_index;
-		}
-
-		he_cl = rb_entry(nd_cl, struct hist_entry, rb_node);
-		c2c_he_cl = container_of(he_cl, struct c2c_hist_entry, he);
-
-		c2c.cacheline_index[index].he_cl = he_cl;
-		c2c.cacheline_index[index].c2c_he_cl = c2c_he_cl;
-		c2c.cacheline_index[index].symbol_accesses = NULL;
-
-		/* Build symbol access list for this cacheline with proper aggregation */
-		if (c2c_he_cl->hists && c2c_he_cl->hists->hists.entries.rb_root.rb_node) {
-			struct rb_node *nd_d = rb_first_cached(&c2c_he_cl->hists->hists.entries);
-
-			while (nd_d) {
-				struct hist_entry *he_d = rb_entry(nd_d, struct hist_entry, rb_node);
-				struct c2c_hist_entry *c2c_he_d = container_of(he_d, struct c2c_hist_entry, he);
-
-				if (he_d->ms.sym && !he_d->filtered) {
-					uint64_t iaddr_d = he_d->mem_info ? mem_info__iaddr(he_d->mem_info)->addr : he_d->ms.sym->start;
-					struct symbol_access *cur = c2c.cacheline_index[index].symbol_accesses;
-					bool merged = false;
-
-					/* Check if we already have an entry for this symbol+iaddr combination */
-					while (cur) {
-						if (cur->sym == he_d->ms.sym && cur->iaddr == iaddr_d) {
-							/* Aggregate statistics */
-							c2c_add_stats(&cur->stats, &c2c_he_d->stats);
-							c2c_add_cstats(&cur->cstats, &c2c_he_d->cstats);
-							merged = true;
-							break;
-						}
-						cur = cur->next;
-					}
-
-					/* Create new entry if not found */
-					if (!merged) {
-						struct symbol_access *sa = malloc(sizeof(struct symbol_access));
-
-						if (sa) {
-							sa->sym = he_d->ms.sym;
-							sa->iaddr = iaddr_d;
-							sa->map = he_d->ms.map;
-							sa->maps = he_d->ms.maps;
-							memcpy(&sa->stats, &c2c_he_d->stats, sizeof(sa->stats));
-							memcpy(&sa->cstats, &c2c_he_d->cstats, sizeof(sa->cstats));
-							sa->next = c2c.cacheline_index[index].symbol_accesses;
-							c2c.cacheline_index[index].symbol_accesses = sa;
-						}
-					}
-				}
-				nd_d = rb_next(nd_d);
-			}
-		}
-
-		index++;
-		c2c.cacheline_index_size = index;  /* Update size as we go */
-		nd_cl = rb_next(nd_cl);
-	}
-
-	/* Mark as built to prevent redundant calls */
-	c2c.cacheline_index_built = true;
-}
-
-/*
- * Optimized helper function to populate cacheline grandchildren for a symbol child entry
- * Returns the number of grandchildren created
- */
-static int populate_cacheline_grandchildren(struct hist_entry *parent_he,
-					   struct hist_entry *child_he,
-					   struct related_symbol *rel_sym)
-{
-	struct rb_root_cached *groot = &child_he->hroot_out;
-	/* temp array to sort grandchildren by Stores */
-	struct grand_item *items = NULL;
-
-	int items_cnt = 0, items_cap = 0;
-	u64 parent_iaddr = parent_he->mem_info ? mem_info__iaddr(parent_he->mem_info)->addr :
-			   (parent_he->ms.sym ? parent_he->ms.sym->start : 0);
-
-	/* Use the pre-built index instead of traversing all cachelines */
-	for (int i = 0; i < c2c.cacheline_index_size; i++) {
-		struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[i];
-		struct symbol_access *parent_access = NULL, *child_access = NULL;
-
-		/* Check if both parent and child symbols access this cacheline */
-		for (struct symbol_access *sa = cl_entry->symbol_accesses; sa; sa = sa->next) {
-			if (symbol_name_equal(sa->sym, parent_he->ms.sym) && sa->iaddr == parent_iaddr)
-				parent_access = sa;
-			else if (symbol_name_equal(sa->sym, rel_sym->sym) && sa->iaddr == rel_sym->iaddr)
-				child_access = sa;
-
-			/* Early exit if both found */
-			if (parent_access && child_access)
-				break;
-		}
-
-		if (parent_access && child_access) {
-			struct c2c_hist_entry *grand_c2c = zalloc(sizeof(*grand_c2c));
-			struct hist_entry *grand_he;
-			u64 child_stores = child_access->stats.store;
-
-			if (!grand_c2c)
-				break;
-
-			init_c2c_he_related_symbols(grand_c2c);
-			grand_he = &grand_c2c->he;
-			/* copy ms from cacheline entry, but clear sym to print cacheline address */
-			memcpy(&grand_he->ms, &cl_entry->he_cl->ms, sizeof(struct map_symbol));
-			grand_he->ms.sym = NULL;
-			grand_he->mem_info = mem_info__get(cl_entry->he_cl->mem_info);
-			grand_he->thread = cl_entry->he_cl->thread;
-			grand_he->cpumode = cl_entry->he_cl->cpumode;
-			grand_he->cpu = cl_entry->he_cl->cpu;
-			grand_he->socket = cl_entry->he_cl->socket;
-			grand_he->parent_he = child_he;
-			grand_he->depth = child_he->depth + 1;
-			grand_he->leaf = true;
-			grand_he->hists = &c2c.symbol_hists.hists;
-			grand_he->filtered = false;
-			grand_he->unfolded = false;
-			grand_he->has_children = false;
-			grand_he->nr_rows = 0;
-			grand_he->row_offset = 0;
-			memset(&grand_he->stat, 0, sizeof(grand_he->stat));
-			grand_he->hroot_in = RB_ROOT_CACHED;
-			grand_he->hroot_out = RB_ROOT_CACHED;
-			INIT_LIST_HEAD(&grand_he->pairs.node);
-
-			/* Initialize hierarchy pointers for grandchild */
-			grand_he->hpp_list = &c2c.symbol_hists.list;
-
-			/* Use pre-computed stats from index - eliminates redundant traversal */
-			memcpy(&grand_c2c->stats, &child_access->stats, sizeof(grand_c2c->stats));
-			memcpy(&grand_c2c->cstats, &child_access->cstats, sizeof(grand_c2c->cstats));
-			/* stats for columns alignment */
-			grand_he->stat.nr_events = grand_c2c->stats.lcl_hitm + grand_c2c->stats.rmt_hitm +
-						grand_c2c->stats.lcl_peer + grand_c2c->stats.rmt_peer;
-			grand_he->stat.period = grand_he->stat.nr_events;
-			grand_he->stat.weight1 = grand_c2c->stats.rmt_hitm + grand_c2c->stats.lcl_hitm;
-
-			/* Initialize stat_acc for grandchild if needed */
-			if (symbol_conf.cumulate_callchain) {
-				grand_he->stat_acc = calloc(1, sizeof(struct he_stat));
-				if (grand_he->stat_acc)
-					memcpy(grand_he->stat_acc, &grand_he->stat, sizeof(struct he_stat));
-			}
-
-			/* push into temp array */
-			if (items_cnt == items_cap) {
-				int new_cap = items_cap ? items_cap * 2 : 8;
-				struct grand_item *ni = realloc(items, new_cap * sizeof(*items));
-
-				if (!ni) {
-					free(grand_c2c);
-					/* Free all previously allocated grand_c2c structures */
-					for (int j = 0; j < items_cnt; j++)
-						free(items[j].grand_c2c);
-					free(items);
-					break;
-				}
-				items = ni; items_cap = new_cap;
-			}
-			items[items_cnt].grand_c2c = grand_c2c;
-			items[items_cnt].grand_he = grand_he;
-			items[items_cnt].stores = child_stores;
-			items_cnt++;
-		}
-	}
-
-	/* sort by stores desc using qsort */
-	if (items_cnt > 0)
-		qsort(items, items_cnt, sizeof(struct grand_item), grand_item_cmp);
-
-	/* insert in order */
-	for (int a = 0; a < items_cnt; a++) {
-		struct rb_node **p = &groot->rb_root.rb_node;
-		struct rb_node *parent = NULL;
-		bool leftmost = true;
-
-		while (*p != NULL) {
-			parent = *p;
-			p = &parent->rb_right;
-			leftmost = false;
-		}
-		rb_link_node(&items[a].grand_he->rb_node, parent, p);
-		rb_insert_color_cached(&items[a].grand_he->rb_node, groot, leftmost);
-	}
-
-	/* mark child has_children only if any grandchildren were added */
-	child_he->has_children = items_cnt > 0;
-	if (items_cnt > 0) {
-		/* Update nr_rows to include grandchildren count */
-		child_he->nr_rows = items_cnt;
-	}
-
-	free(items);
-	return items_cnt;
-}
-
-void populate_symbol_children(struct hist_entry *he)
-{
-	struct c2c_hist_entry *c2c_he;
-	struct related_symbol **sorted;
-	struct rb_root_cached *root;
-	struct rb_node **p, *parent;
-	bool leftmost;
-	int count = 0, num_rel, i;
-
-	/* Validate and prepare entries */
-	c2c_he = validate_and_prepare_entries(he);
-	if (!c2c_he)
-		return;
-
-	root = &he->hroot_out;
-
-	/* Sort related symbols by stores descending */
-	sorted = sort_related_symbols_by_stores(c2c_he, &num_rel);
-	if (!sorted)
-		return;
-
-	/* Create and populate child entries */
-	for (i = 0; i < num_rel; i++) {
-		struct hist_entry *child_he;
-		struct related_symbol *rel_sym = sorted[i];
-
-		/* Create child entry */
-		child_he = create_symbol_child_entry(he, rel_sym);
-		if (!child_he)
-			continue;
-
-		/* Populate grandchildren (cachelines) */
-		populate_cacheline_grandchildren(he, child_he, rel_sym);
-
-		/* Insert child into the parent's tree */
-		p = &root->rb_root.rb_node;
-		parent = NULL;
-		leftmost = true;
-		while (*p != NULL) {
-			parent = *p;
-			p = &parent->rb_right;
-			leftmost = false;
-		}
-
-		rb_link_node(&child_he->rb_node, parent, p);
-		rb_insert_color_cached(&child_he->rb_node, root, leftmost);
-
-		count++;
-	}
-
-	free(sorted);
-
-	/* Update parent's nr_rows and has_children flag */
-	if (count > 0) {
-		he->nr_rows = count;
-		he->has_children = true;
-	} else {
-		he->nr_rows = 0;
-		he->has_children = false;
-	}
-}
-
-/*
- * Build symbol associations based on cacheline sharing
- *
- * Logic: When multiple symbols access the same cacheline (false sharing),
- * they are considered related. The association count increases by 1 for
- * each shared cacheline between two symbols.
- */
-static void build_symbol_associations(void)
-{
-	struct rb_node *nd_sym;
-	struct hist_entry *he_sym;
-	struct c2c_hist_entry *c2c_he_sym;
-	int cl_idx;
-
-	/*
-	 * Algorithm:
-	 * 1. For each cacheline that has HITM events
-	 * 2. Look at its detailed access records
-	 * 3. Find all symbols that accessed it with HITM
-	 * 4. Create associations between these symbols
-	 */
-
-	/* Phase 1: Use cached index to find symbol conflicts efficiently */
-	/* Build cacheline index (will only build once due to internal flag) */
-	build_cacheline_symbol_index();
-
-	/* Iterate through cached index instead of rb-tree */
-	for (cl_idx = 0; cl_idx < c2c.cacheline_index_size; cl_idx++) {
-		struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[cl_idx];
-		struct c2c_hist_entry *c2c_he_cl = cl_entry->c2c_he_cl;
-		struct symbol_addr_pair {
-			struct symbol *sym;
-			uint64_t iaddr;
-		} *symbols_with_hitm = NULL;
-		int symbol_count = 0;
-		int symbol_capacity = 0;
-		struct symbol_access *sa;
-		int i, j;
-
-		/* Skip cachelines without HITM events */
-		if ((c2c_he_cl->stats.rmt_hitm + c2c_he_cl->stats.lcl_hitm) == 0)
-			continue;
-
-		/* Collect all (symbol, address) pairs that accessed this cacheline with HITM */
-		for (sa = cl_entry->symbol_accesses; sa; sa = sa->next) {
-			if (sa->sym && (sa->stats.rmt_hitm + sa->stats.lcl_hitm) > 0) {
-				uint64_t iaddr = sa->iaddr ? sa->iaddr : sa->sym->start;
-				/* Add (symbol, iaddr) pair to list if not already there */
-				bool found = false;
-
-				for (i = 0; i < symbol_count; i++) {
-					if (symbol_name_equal(symbols_with_hitm[i].sym, sa->sym) &&
-					    symbols_with_hitm[i].iaddr == iaddr) {
-						found = true;
-						break;
-					}
-				}
-
-				if (!found) {
-					if (symbol_count >= symbol_capacity) {
-						struct symbol_addr_pair *new_symbols;
-
-						symbol_capacity = symbol_capacity ? symbol_capacity * 2 : 4;
-						new_symbols = realloc(symbols_with_hitm,
-								    symbol_capacity * sizeof(struct symbol_addr_pair));
-						if (!new_symbols) {
-							/* Memory allocation failed, skip this symbol */
-							continue;
-						}
-						symbols_with_hitm = new_symbols;
-					}
-					if (symbols_with_hitm) {
-						symbols_with_hitm[symbol_count].sym = sa->sym;
-						symbols_with_hitm[symbol_count].iaddr = iaddr;
-						symbol_count++;
-					}
-				}
-			}
-		}
-
-		/* Create associations between all symbols that conflict on this cacheline */
-		if (symbol_count > 1) {
-			for (i = 0; i < symbol_count; i++) {
-				/* Find symbol entry in symbol_hists */
-				nd_sym = rb_first_cached(&c2c.symbol_hists.hists.entries);
-				while (nd_sym) {
-					he_sym = rb_entry(nd_sym, struct hist_entry, rb_node);
-					/* Match parent entry by BOTH symbol and code address */
-					{
-						uint64_t parent_iaddr = 0;
-
-						if (he_sym->mem_info)
-							parent_iaddr = mem_info__iaddr(he_sym->mem_info)->addr;
-						else if (he_sym->ms.sym)
-							parent_iaddr = he_sym->ms.sym->start;
-
-						if (symbol_name_equal(he_sym->ms.sym, symbols_with_hitm[i].sym) &&
-							parent_iaddr == symbols_with_hitm[i].iaddr) {
-							c2c_he_sym = container_of(he_sym, struct c2c_hist_entry, he);
-
-							/* Add all other symbols as related */
-							for (j = 0; j < symbol_count; j++) {
-								if (i != j) {
-									struct related_symbol *rel_sym;
-									bool exists = false;
-
-									/* Check if already added (compare both sym and iaddr) */
-									list_for_each_entry(rel_sym, &c2c_he_sym->related_symbols, list) {
-										if (symbol_name_equal(rel_sym->sym, symbols_with_hitm[j].sym) &&
-											rel_sym->iaddr == symbols_with_hitm[j].iaddr) {
-											exists = true;
-											break;
-										}
-									}
-
-									if (!exists) {
-										rel_sym = zalloc(sizeof(*rel_sym));
-										if (rel_sym) {
-											rel_sym->sym = symbols_with_hitm[j].sym;
-											rel_sym->iaddr = symbols_with_hitm[j].iaddr;
-											/* zalloc already zeros memory, no need for memset */
-											list_add_tail(&rel_sym->list, &c2c_he_sym->related_symbols);
-										}
-									}
-								}
-							}
-
-							/* Mark as having children */
-							if (!list_empty(&c2c_he_sym->related_symbols)) {
-								he_sym->has_children = true;
-								he_sym->unfolded = false;
-								he_sym->leaf = false;
-							}
-
-							break;
-						}
-					}
-					nd_sym = rb_next(nd_sym);
-				}
-			}
-		}
-
-		free(symbols_with_hitm);
-	}
-
-	/* Phase 2: Aggregate stats for related symbols using cached index */
-	nd_sym = rb_first_cached(&c2c.symbol_hists.hists.entries);
-	while (nd_sym) {
-		struct related_symbol *rel_sym;
-
-		he_sym = rb_entry(nd_sym, struct hist_entry, rb_node);
-		c2c_he_sym = container_of(he_sym, struct c2c_hist_entry, he);
-
-		/* For each related symbol, aggregate stats from shared cachelines using cached index */
-		list_for_each_entry(rel_sym, &c2c_he_sym->related_symbols, list) {
-			/* Precompute parent's iaddr once */
-			uint64_t parent_iaddr = he_sym->mem_info ?
-				mem_info__iaddr(he_sym->mem_info)->addr :
-				(he_sym->ms.sym ? he_sym->ms.sym->start : 0);
-
-		/* Use cached index instead of nested rb_next loops */
-		for (cl_idx = 0; cl_idx < c2c.cacheline_index_size; cl_idx++) {
-			struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[cl_idx];
-				struct symbol_access *sa;
-				bool target_found = false;
-
-				/* First pass: check if target symbol accessed this cacheline */
-				for (sa = cl_entry->symbol_accesses; sa; sa = sa->next) {
-					if (symbol_name_equal(sa->sym, he_sym->ms.sym) && sa->iaddr == parent_iaddr) {
-						target_found = true;
-						break;
-					}
-				}
-
-				if (!target_found)
-					continue;
-
-				/* Second pass: aggregate related symbol stats */
-				for (sa = cl_entry->symbol_accesses; sa; sa = sa->next) {
-					if (symbol_name_equal(sa->sym, rel_sym->sym) && sa->iaddr == rel_sym->iaddr) {
-						c2c_add_stats(&rel_sym->stats, &sa->stats);
-						c2c_add_cstats(&rel_sym->cstats, &sa->cstats);
-					}
-				}
-			}
-		}
-
-		nd_sym = rb_next(nd_sym);
-	}
-
-	/* Phase 3: Create child entries for symbols with associations */
-	nd_sym = rb_first_cached(&c2c.symbol_hists.hists.entries);
-	while (nd_sym) {
-		he_sym = rb_entry(nd_sym, struct hist_entry, rb_node);
-		if (he_sym->has_children)
-			populate_symbol_children(he_sym);
-		nd_sym = rb_next(nd_sym);
-	}
-}
-
-/* Cleanup function for the cacheline symbol index */
-static void cleanup_cacheline_symbol_index(void)
-{
-	if (c2c.cacheline_index) {
-		for (int i = 0; i < c2c.cacheline_index_size; i++) {
-			struct symbol_access *sa = c2c.cacheline_index[i].symbol_accesses;
-
-			while (sa) {
-				struct symbol_access *next = sa->next;
-
-				free(sa);
-				sa = next;
-			}
-		}
-		free(c2c.cacheline_index);
-		c2c.cacheline_index = NULL;
-		c2c.cacheline_index_size = 0;
-		c2c.cacheline_index_capacity = 0;
-		c2c.cacheline_index_built = false;
-	}
-}
-
-static int build_symbol_hists(struct perf_env *env)
-{
-	struct rb_node *next;
-	struct hist_entry *he_sym;
-	struct c2c_hist_entry *c2c_he_sym;
-	struct addr_location al;
-	struct perf_sample sample = {};
-	struct thread *synthetic_thread = NULL;
-	int ret, i;
-
-	/* Invalidate cached total cycles before (re)building symbol histograms */
-	c2c.symbol_total_cycles_valid = false;
-	c2c.symbol_total_cycles = 0;
-
-	next = rb_first_cached(&c2c.hists.hists.entries);
-
-	/* Initialize symbol hists with sort by iaddr (code address) and symbol */
-	ret = c2c_hists__init(&c2c.symbol_hists, "iaddr,symbol", 2, env);
-	if (ret)
-		return ret;
-
-	/* Setup output fields for symbol view - sorted by cycles percentage (descending) */
-	ret = c2c_hists__reinit(&c2c.symbol_hists,
-		"cycles_percent,total_stores,iaddr,symbol,cacheline_symbol",
-		"cycles_percent", env);
-	if (ret)
-		return ret;
-
-	/* Get first thread for consistent aggregation */
-	if (next) {
-		struct hist_entry *first_he = rb_entry(next, struct hist_entry, rb_node);
-
-		synthetic_thread = first_he->thread;
-	}
-
-	/* Build cacheline index (will only build once due to internal flag) */
-	build_cacheline_symbol_index();
-
-	if (c2c.cacheline_index_size == 0) {
-		ui__error("Cacheline index is empty, cannot build symbol hists\n");
-		return -1;
-	}
-
-	/* Directly create histogram entries from cached symbol_access data */
-	for (i = 0; i < c2c.cacheline_index_size; i++) {
-		struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[i];
-		struct symbol_access *sa = cl_entry->symbol_accesses;
-
-		/* Process all symbol accesses for this cacheline */
-		while (sa) {
-			if (sa->sym) {
-				/* Create mem_info with proper instruction address for display */
-				struct mem_info *mi_display = mem_info__new();
-
-				if (mi_display) {
-					mem_info__iaddr(mi_display)->addr = sa->iaddr;
-					mem_info__iaddr(mi_display)->ms.maps = sa->maps;
-					mem_info__iaddr(mi_display)->ms.map = sa->map;
-					mem_info__iaddr(mi_display)->ms.sym = sa->sym;
-					/* Set data address to 0 for consistent display */
-					mem_info__daddr(mi_display)->addr = 0;
-				}
-
-				/* Create consistent address location for symbol aggregation */
-				addr_location__init(&al);
-				al.thread = synthetic_thread;
-				al.maps = sa->maps;
-				al.map = sa->map;
-				al.sym = sa->sym;
-				al.addr = sa->iaddr;
-				al.level = PERF_RECORD_MISC_KERNEL;
-				al.cpumode = PERF_RECORD_MISC_KERNEL;
-				al.cpu = 0;
-				al.socket = 0;
-				al.filtered = 0;
-
-				/* Create sample with consistent values */
-				sample.period = 1;
-				sample.weight = 1;
-				sample.ip = sa->iaddr;
-				sample.pid = synthetic_thread ? thread__pid(synthetic_thread) : 0;
-				sample.tid = synthetic_thread ? thread__tid(synthetic_thread) : 0;
-				sample.cpu = 0;
-				sample.time = 0;
-				sample.addr = 0;
-				sample.id = 0;
-
-				/* Add entry to histogram with mem_info for proper address display */
-				he_sym = hists__add_entry_ops(&c2c.symbol_hists.hists,
-							      &c2c_entry_ops,
-							      &al, NULL, NULL, mi_display,
-							      NULL, &sample, true);
-
-				addr_location__exit(&al);
-				if (mi_display)
-					mem_info__put(mi_display);
-
-				if (he_sym) {
-					c2c_he_sym = container_of(he_sym, struct c2c_hist_entry, he);
-
-					/* Copy aggregated stats from cached symbol_access */
-					c2c_he_sym->stats = sa->stats;
-					c2c_he_sym->cstats = sa->cstats;
-					c2c_add_stats(&c2c.symbol_hists.stats, &sa->stats);
-
-					hists__inc_nr_samples(&c2c.symbol_hists.hists, he_sym->filtered);
-					he_sym->hpp_list = &c2c.symbol_hists.list;
-				}
-			}
-			sa = sa->next;
-		}
-	}
-
-	/* Resort symbol hists */
-	hists__collapse_resort(&c2c.symbol_hists.hists, NULL);
-	hists__output_resort(&c2c.symbol_hists.hists, NULL);
-
-	/* Enable hierarchy support for symbol view to allow multi-level display */
-	c2c.symbol_hists.hists.symbol_filter_str = NULL;
-	c2c.symbol_hists.hists.socket_filter = -1;
-
-	/* Initialize the hist browser fields needed for hierarchy */
-	c2c.symbol_hists.hists.nr_hpp_node = 0;
-
-	/* Build symbol associations after hists are complete */
-	build_symbol_associations();
-
-	/* Note: We no longer clean up cacheline_index here to avoid rebuilding
-	 * it multiple times when called from different contexts (stdio/browse).
-	 * The index will be automatically cleaned up when the c2c structure
-	 * goes out of scope or can be explicitly cleaned at program exit.
-	 */
-
-	/* Precompute and cache total cycles to speed up percent rendering */
-	(void)get_total_cycles_all_symbols();
-
-	return 0;
-}
 
 static void c2c_browser__update_nr_entries(struct hist_browser *hb)
 {
