+++ tools/perf/ui/browsers/c2c-symbol.c
@@ -2,9 +2,21 @@
 /**
  * C2C Symbol Browser - Display symbol-level cacheline sharing analysis
  * 
- * This browser provides a symbol-centric view of cache-to-cache (C2C) data,
- * showing which symbols share cachelines and may experience false sharing.
+ * This file contains the symbol view implementation for C2C analysis,
+ * including:
+ * - Symbol histogram building and management
+ * - Cacheline-to-symbol index for efficient lookups
+ * - Symbol association building (finding related symbols)
+ * - Symbol browser UI components
  */
+
+#include <stdlib.h>
+#include <string.h>
+#include <linux/zalloc.h>
+#include <linux/list.h>
+#include <linux/rbtree.h>
+#include <linux/string.h>
+#include <sys/ttydefaults.h>
 
 #include "../../builtin-c2c.h"
 #include "../browser.h"
@@ -16,13 +28,1035 @@
 #include "../../util/mem-info.h"
 #include "../../util/cacheline.h"
 #include "../../util/debug.h"
-#include <linux/zalloc.h>
-#include <linux/string.h>
-#include <sys/ttydefaults.h>
+#include "../../util/thread.h"
+#include "../../util/addr_location.h"
+#include "../../util/event.h"
 
+/*
+ * ============================================================================
+ * Internal structures for symbol view processing
+ * ============================================================================
+ */
 
-struct c2c_hist_entry;
+/**
+ * struct grand_item - Grandchild entry for sorting
+ * @grand_c2c: C2C histogram entry for the grandchild
+ * @grand_he: Histogram entry for the grandchild
+ * @stores: Store count for sorting
+ */
+struct grand_item {
+	struct c2c_hist_entry	*grand_c2c;
+	struct hist_entry	*grand_he;
+	u64			 stores;
+};
 
+/*
+ * ============================================================================
+ * Comparison functions for qsort
+ * ============================================================================
+ */
+
+/**
+ * related_symbol_cmp - Compare related symbols by store count (descending)
+ */
+static int related_symbol_cmp(const void *a, const void *b)
+{
+	const struct related_symbol *sym_a = *(const struct related_symbol **)a;
+	const struct related_symbol *sym_b = *(const struct related_symbol **)b;
+
+	if (sym_b->stats.store > sym_a->stats.store)
+		return 1;
+	else if (sym_b->stats.store < sym_a->stats.store)
+		return -1;
+	return 0;
+}
+
+/**
+ * grand_item_cmp - Compare grand items by store count (descending)
+ */
+static int grand_item_cmp(const void *a, const void *b)
+{
+	const struct grand_item *item_a = (const struct grand_item *)a;
+	const struct grand_item *item_b = (const struct grand_item *)b;
+
+	if (item_b->stores > item_a->stores)
+		return 1;
+	else if (item_b->stores < item_a->stores)
+		return -1;
+	return 0;
+}
+
+/*
+ * ============================================================================
+ * Helper functions for symbol view processing
+ * ============================================================================
+ */
+
+/**
+ * validate_and_prepare_entries - Validate and prepare for populating symbol children
+ * @he: Histogram entry to validate
+ *
+ * Returns the c2c_hist_entry if valid, NULL otherwise
+ */
+static struct c2c_hist_entry *validate_and_prepare_entries(struct hist_entry *he)
+{
+	struct c2c_hist_entry *c2c_he;
+	struct rb_root_cached *root;
+
+	if (!he || !he->has_children)
+		return NULL;
+
+	root = &he->hroot_out;
+
+	/* If already populated, return */
+	if (!RB_EMPTY_ROOT(&root->rb_root))
+		return NULL;
+
+	c2c_he = container_of(he, struct c2c_hist_entry, he);
+	if (!c2c_he)
+		return NULL;
+
+	/* Ensure related_symbols list is valid */
+	if (list_empty(&c2c_he->related_symbols)) {
+		he->has_children = false;  /* Reset inconsistent state */
+		return NULL;
+	}
+
+	return c2c_he;
+}
+
+/**
+ * sort_related_symbols_by_stores - Sort related symbols by stores descending
+ * @c2c_he: C2C histogram entry containing related symbols
+ * @num_rel_out: Output parameter for number of related symbols
+ *
+ * Returns allocated array of sorted related_symbol pointers, or NULL on error.
+ * Caller must free the returned array.
+ */
+static struct related_symbol **sort_related_symbols_by_stores(struct c2c_hist_entry *c2c_he,
+							      int *num_rel_out)
+{
+	struct related_symbol *rel_sym;
+	struct related_symbol **sorted = NULL;
+	int num_rel = 0, idx = 0;
+
+	/* Count related symbols and allocate array in one pass */
+	list_for_each_entry(rel_sym, &c2c_he->related_symbols, list)
+		num_rel++;
+
+	if (num_rel == 0) {
+		*num_rel_out = 0;
+		return NULL;
+	}
+
+	sorted = calloc(num_rel, sizeof(*sorted));
+	if (!sorted) {
+		*num_rel_out = 0;
+		return NULL;
+	}
+
+	/* Fill array in the same pass */
+	list_for_each_entry(rel_sym, &c2c_he->related_symbols, list)
+		sorted[idx++] = rel_sym;
+
+	/* Sort by stats.store descending */
+	qsort(sorted, num_rel, sizeof(*sorted), related_symbol_cmp);
+
+	*num_rel_out = num_rel;
+	return sorted;
+}
+
+/**
+ * create_symbol_child_entry - Create and initialize a symbol child entry
+ * @parent_he: Parent histogram entry
+ * @rel_sym: Related symbol to create entry for
+ *
+ * Returns the created hist_entry or NULL on error
+ */
+static struct hist_entry *create_symbol_child_entry(struct hist_entry *parent_he,
+						    struct related_symbol *rel_sym)
+{
+	struct c2c_hist_entry *child_c2c_he, *child_c2c;
+	struct hist_entry *child_he;
+
+	if (!rel_sym || !rel_sym->sym)
+		return NULL;
+
+	/* Allocate child hist_entry - simplified version for symbol children */
+	child_c2c_he = zalloc(sizeof(*child_c2c_he));
+	if (!child_c2c_he)
+		return NULL;
+
+	/* Initialize the related_symbols list */
+	init_c2c_he_related_symbols(child_c2c_he);
+
+	child_he = &child_c2c_he->he;
+
+	/* Complete initialization - copy parent's map_symbol structure first */
+	memcpy(&child_he->ms, &parent_he->ms, sizeof(struct map_symbol));
+	/* Then override the symbol and address */
+	child_he->ms.sym = rel_sym->sym;
+
+	/* Create a synthetic mem_info to store the iaddr for proper display */
+	if (parent_he->mem_info) {
+		child_he->mem_info = mem_info__clone(parent_he->mem_info);
+		if (child_he->mem_info) {
+			/* Set the instruction address to the related symbol's iaddr */
+			mem_info__iaddr(child_he->mem_info)->addr = rel_sym->iaddr;
+		}
+	}
+
+	/* Copy essential fields from parent */
+	child_he->thread = parent_he->thread;
+	child_he->cpumode = parent_he->cpumode;
+	child_he->cpu = parent_he->cpu;
+	child_he->socket = parent_he->socket;
+
+	/* Set hierarchy fields */
+	child_he->parent_he = parent_he;
+	child_he->depth = parent_he->depth + 1;
+	child_he->leaf = false; /* it can have grandchildren (cachelines) */
+	/* In symbol view, all entries should use symbol_hists */
+	child_he->hists = &c2c.symbol_hists.hists;
+	child_he->filtered = false;  /* Make sure it's not filtered out */
+	child_he->unfolded = false;
+	child_he->has_children = false; /* Will be set to true only if grandchildren are added */
+	child_he->has_no_entry = false;
+	child_he->nr_rows = 0;
+	child_he->row_offset = 0;
+
+	/* Initialize stats properly */
+	memset(&child_he->stat, 0, sizeof(child_he->stat));
+
+	/* Set stat values based on c2c stats */
+	child_he->stat.nr_events = rel_sym->stats.rmt_hitm + rel_sym->stats.lcl_hitm +
+				   rel_sym->stats.rmt_peer + rel_sym->stats.lcl_peer;
+	child_he->stat.period = child_he->stat.nr_events;
+
+	/* These weight fields are used by some columns */
+	child_he->stat.weight1 = rel_sym->stats.rmt_hitm + rel_sym->stats.lcl_hitm;
+
+	/* Initialize stat_acc - allocate if needed */
+	if (symbol_conf.cumulate_callchain) {
+		child_he->stat_acc = calloc(1, sizeof(struct he_stat));
+		if (child_he->stat_acc)
+			memcpy(child_he->stat_acc, &child_he->stat, sizeof(struct he_stat));
+	}
+
+	/* Initialize rb trees */
+	child_he->hroot_in = RB_ROOT_CACHED;
+	child_he->hroot_out = RB_ROOT_CACHED;
+
+	/* Initialize pairs list */
+	INIT_LIST_HEAD(&child_he->pairs.node);
+
+	/* Initialize hierarchy pointers */
+	child_he->hpp_list = &c2c.symbol_hists.list;
+
+	/* Copy c2c stats - this is what c2c columns use */
+	child_c2c = container_of(child_he, struct c2c_hist_entry, he);
+	memcpy(&child_c2c->stats, &rel_sym->stats, sizeof(rel_sym->stats));
+	memcpy(&child_c2c->cstats, &rel_sym->cstats, sizeof(rel_sym->cstats));
+	init_c2c_he_related_symbols(child_c2c);
+
+	/* Build cacheline grandchildren under each related symbol child */
+	child_c2c->hists = NULL;
+
+	return child_he;
+}
+
+/*
+ * ============================================================================
+ * Cacheline Symbol Index Management
+ * ============================================================================
+ */
+
+/**
+ * build_cacheline_symbol_index - Build index mapping cachelines to accessing symbols
+ *
+ * Creates an optimized index structure for looking up which symbols
+ * access each cacheline. This is used for building symbol associations
+ * and populating child entries efficiently.
+ */
+void build_cacheline_symbol_index(void)
+{
+	struct rb_node *nd_cl;
+	int index = 0;
+
+	/* Return early if already built */
+	if (c2c.cacheline_index_built)
+		return;
+
+	/* Free existing index if any */
+	if (c2c.cacheline_index) {
+		for (int i = 0; i < c2c.cacheline_index_size; i++) {
+			struct symbol_access *sa = c2c.cacheline_index[i].symbol_accesses;
+
+			while (sa) {
+				struct symbol_access *next = sa->next;
+
+				free(sa);
+				sa = next;
+			}
+		}
+		free(c2c.cacheline_index);
+		c2c.cacheline_index = NULL;
+	}
+
+	/* Build index in single pass with dynamic array growth */
+	c2c.cacheline_index_size = 0;
+	c2c.cacheline_index_capacity = 64; /* Start with reasonable size */
+
+	c2c.cacheline_index = malloc(c2c.cacheline_index_capacity * sizeof(struct cacheline_symbol_entry));
+	if (!c2c.cacheline_index) {
+		c2c.cacheline_index_size = 0;
+		return;
+	}
+
+	nd_cl = rb_first_cached(&c2c.hists.hists.entries);
+	while (nd_cl) {
+		struct hist_entry *he_cl;
+		struct c2c_hist_entry *c2c_he_cl;
+
+		/* Grow array if needed */
+		if (index >= c2c.cacheline_index_capacity) {
+			struct cacheline_symbol_entry *new_index;
+			int cleanup_i;
+
+			c2c.cacheline_index_capacity *= 2;
+			new_index = realloc(c2c.cacheline_index,
+				c2c.cacheline_index_capacity * sizeof(struct cacheline_symbol_entry));
+			if (!new_index) {
+				/* Cleanup on allocation failure */
+				for (cleanup_i = 0; cleanup_i < c2c.cacheline_index_size; cleanup_i++) {
+					struct symbol_access *sa = c2c.cacheline_index[cleanup_i].symbol_accesses;
+
+					while (sa) {
+						struct symbol_access *next = sa->next;
+
+						free(sa);
+						sa = next;
+					}
+				}
+				free(c2c.cacheline_index);
+				c2c.cacheline_index = NULL;
+				c2c.cacheline_index_size = 0;
+				return;
+			}
+			c2c.cacheline_index = new_index;
+		}
+
+		he_cl = rb_entry(nd_cl, struct hist_entry, rb_node);
+		c2c_he_cl = container_of(he_cl, struct c2c_hist_entry, he);
+
+		c2c.cacheline_index[index].he_cl = he_cl;
+		c2c.cacheline_index[index].c2c_he_cl = c2c_he_cl;
+		c2c.cacheline_index[index].symbol_accesses = NULL;
+
+		/* Build symbol access list for this cacheline with proper aggregation */
+		if (c2c_he_cl->hists && c2c_he_cl->hists->hists.entries.rb_root.rb_node) {
+			struct rb_node *nd_d = rb_first_cached(&c2c_he_cl->hists->hists.entries);
+
+			while (nd_d) {
+				struct hist_entry *he_d = rb_entry(nd_d, struct hist_entry, rb_node);
+				struct c2c_hist_entry *c2c_he_d = container_of(he_d, struct c2c_hist_entry, he);
+
+				if (he_d->ms.sym && !he_d->filtered) {
+					uint64_t iaddr_d = he_d->mem_info ? mem_info__iaddr(he_d->mem_info)->addr : he_d->ms.sym->start;
+					struct symbol_access *cur = c2c.cacheline_index[index].symbol_accesses;
+					bool merged = false;
+
+					/* Check if we already have an entry for this symbol+iaddr combination */
+					while (cur) {
+						if (cur->sym == he_d->ms.sym && cur->iaddr == iaddr_d) {
+							/* Aggregate statistics */
+							c2c_add_stats(&cur->stats, &c2c_he_d->stats);
+							c2c_add_cstats(&cur->cstats, &c2c_he_d->cstats);
+							merged = true;
+							break;
+						}
+						cur = cur->next;
+					}
+
+					/* Create new entry if not found */
+					if (!merged) {
+						struct symbol_access *sa = malloc(sizeof(struct symbol_access));
+
+						if (sa) {
+							sa->sym = he_d->ms.sym;
+							sa->iaddr = iaddr_d;
+							sa->map = he_d->ms.map;
+							sa->maps = he_d->ms.maps;
+							memcpy(&sa->stats, &c2c_he_d->stats, sizeof(sa->stats));
+							memcpy(&sa->cstats, &c2c_he_d->cstats, sizeof(sa->cstats));
+							sa->next = c2c.cacheline_index[index].symbol_accesses;
+							c2c.cacheline_index[index].symbol_accesses = sa;
+						}
+					}
+				}
+				nd_d = rb_next(nd_d);
+			}
+		}
+
+		index++;
+		c2c.cacheline_index_size = index;  /* Update size as we go */
+		nd_cl = rb_next(nd_cl);
+	}
+
+	/* Mark as built to prevent redundant calls */
+	c2c.cacheline_index_built = true;
+}
+
+/**
+ * cleanup_cacheline_symbol_index - Free the cacheline symbol index
+ */
+void cleanup_cacheline_symbol_index(void)
+{
+	if (c2c.cacheline_index) {
+		for (int i = 0; i < c2c.cacheline_index_size; i++) {
+			struct symbol_access *sa = c2c.cacheline_index[i].symbol_accesses;
+
+			while (sa) {
+				struct symbol_access *next = sa->next;
+
+				free(sa);
+				sa = next;
+			}
+		}
+		free(c2c.cacheline_index);
+		c2c.cacheline_index = NULL;
+		c2c.cacheline_index_size = 0;
+		c2c.cacheline_index_capacity = 0;
+		c2c.cacheline_index_built = false;
+	}
+}
+
+/*
+ * ============================================================================
+ * Symbol Children Population
+ * ============================================================================
+ */
+
+/**
+ * populate_cacheline_grandchildren - Populate cacheline grandchildren for a symbol child
+ * @parent_he: Top-level parent histogram entry
+ * @child_he: Child histogram entry (related symbol)
+ * @rel_sym: Related symbol structure
+ *
+ * Returns the number of grandchildren created
+ */
+static int populate_cacheline_grandchildren(struct hist_entry *parent_he,
+					    struct hist_entry *child_he,
+					    struct related_symbol *rel_sym)
+{
+	struct rb_root_cached *groot = &child_he->hroot_out;
+	/* temp array to sort grandchildren by Stores */
+	struct grand_item *items = NULL;
+
+	int items_cnt = 0, items_cap = 0;
+	u64 parent_iaddr = parent_he->mem_info ? mem_info__iaddr(parent_he->mem_info)->addr :
+			   (parent_he->ms.sym ? parent_he->ms.sym->start : 0);
+
+	/* Use the pre-built index instead of traversing all cachelines */
+	for (int i = 0; i < c2c.cacheline_index_size; i++) {
+		struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[i];
+		struct symbol_access *parent_access = NULL, *child_access = NULL;
+
+		/* Check if both parent and child symbols access this cacheline */
+		for (struct symbol_access *sa = cl_entry->symbol_accesses; sa; sa = sa->next) {
+			if (symbol_name_equal(sa->sym, parent_he->ms.sym) && sa->iaddr == parent_iaddr)
+				parent_access = sa;
+			else if (symbol_name_equal(sa->sym, rel_sym->sym) && sa->iaddr == rel_sym->iaddr)
+				child_access = sa;
+
+			/* Early exit if both found */
+			if (parent_access && child_access)
+				break;
+		}
+
+		if (parent_access && child_access) {
+			struct c2c_hist_entry *grand_c2c = zalloc(sizeof(*grand_c2c));
+			struct hist_entry *grand_he;
+			u64 child_stores = child_access->stats.store;
+
+			if (!grand_c2c)
+				break;
+
+			init_c2c_he_related_symbols(grand_c2c);
+			grand_he = &grand_c2c->he;
+			/* copy ms from cacheline entry, but clear sym to print cacheline address */
+			memcpy(&grand_he->ms, &cl_entry->he_cl->ms, sizeof(struct map_symbol));
+			grand_he->ms.sym = NULL;
+			grand_he->mem_info = mem_info__get(cl_entry->he_cl->mem_info);
+			grand_he->thread = cl_entry->he_cl->thread;
+			grand_he->cpumode = cl_entry->he_cl->cpumode;
+			grand_he->cpu = cl_entry->he_cl->cpu;
+			grand_he->socket = cl_entry->he_cl->socket;
+			grand_he->parent_he = child_he;
+			grand_he->depth = child_he->depth + 1;
+			grand_he->leaf = true;
+			grand_he->hists = &c2c.symbol_hists.hists;
+			grand_he->filtered = false;
+			grand_he->unfolded = false;
+			grand_he->has_children = false;
+			grand_he->nr_rows = 0;
+			grand_he->row_offset = 0;
+			memset(&grand_he->stat, 0, sizeof(grand_he->stat));
+			grand_he->hroot_in = RB_ROOT_CACHED;
+			grand_he->hroot_out = RB_ROOT_CACHED;
+			INIT_LIST_HEAD(&grand_he->pairs.node);
+
+			/* Initialize hierarchy pointers for grandchild */
+			grand_he->hpp_list = &c2c.symbol_hists.list;
+
+			/* Use pre-computed stats from index - eliminates redundant traversal */
+			memcpy(&grand_c2c->stats, &child_access->stats, sizeof(grand_c2c->stats));
+			memcpy(&grand_c2c->cstats, &child_access->cstats, sizeof(grand_c2c->cstats));
+			/* stats for columns alignment */
+			grand_he->stat.nr_events = grand_c2c->stats.lcl_hitm + grand_c2c->stats.rmt_hitm +
+						grand_c2c->stats.lcl_peer + grand_c2c->stats.rmt_peer;
+			grand_he->stat.period = grand_he->stat.nr_events;
+			grand_he->stat.weight1 = grand_c2c->stats.rmt_hitm + grand_c2c->stats.lcl_hitm;
+
+			/* Initialize stat_acc for grandchild if needed */
+			if (symbol_conf.cumulate_callchain) {
+				grand_he->stat_acc = calloc(1, sizeof(struct he_stat));
+				if (grand_he->stat_acc)
+					memcpy(grand_he->stat_acc, &grand_he->stat, sizeof(struct he_stat));
+			}
+
+			/* push into temp array */
+			if (items_cnt == items_cap) {
+				int new_cap = items_cap ? items_cap * 2 : 8;
+				struct grand_item *ni = realloc(items, new_cap * sizeof(*items));
+
+				if (!ni) {
+					free(grand_c2c);
+					/* Free all previously allocated grand_c2c structures */
+					for (int j = 0; j < items_cnt; j++)
+						free(items[j].grand_c2c);
+					free(items);
+					break;
+				}
+				items = ni;
+				items_cap = new_cap;
+			}
+			items[items_cnt].grand_c2c = grand_c2c;
+			items[items_cnt].grand_he = grand_he;
+			items[items_cnt].stores = child_stores;
+			items_cnt++;
+		}
+	}
+
+	/* sort by stores desc using qsort */
+	if (items_cnt > 0)
+		qsort(items, items_cnt, sizeof(struct grand_item), grand_item_cmp);
+
+	/* insert in order */
+	for (int a = 0; a < items_cnt; a++) {
+		struct rb_node **p = &groot->rb_root.rb_node;
+		struct rb_node *parent = NULL;
+		bool leftmost = true;
+
+		while (*p != NULL) {
+			parent = *p;
+			p = &parent->rb_right;
+			leftmost = false;
+		}
+		rb_link_node(&items[a].grand_he->rb_node, parent, p);
+		rb_insert_color_cached(&items[a].grand_he->rb_node, groot, leftmost);
+	}
+
+	/* mark child has_children only if any grandchildren were added */
+	child_he->has_children = items_cnt > 0;
+	if (items_cnt > 0) {
+		/* Update nr_rows to include grandchildren count */
+		child_he->nr_rows = items_cnt;
+	}
+
+	free(items);
+	return items_cnt;
+}
+
+/**
+ * populate_symbol_children - Create child entries for a symbol
+ * @he: Parent histogram entry to populate
+ *
+ * Creates child entries (related symbols) under the given parent entry.
+ * Each child represents a symbol that shares a cacheline with the parent.
+ */
+void populate_symbol_children(struct hist_entry *he)
+{
+	struct c2c_hist_entry *c2c_he;
+	struct related_symbol **sorted;
+	struct rb_root_cached *root;
+	struct rb_node **p, *parent;
+	bool leftmost;
+	int count = 0, num_rel, i;
+
+	/* Validate and prepare entries */
+	c2c_he = validate_and_prepare_entries(he);
+	if (!c2c_he)
+		return;
+
+	root = &he->hroot_out;
+
+	/* Sort related symbols by stores descending */
+	sorted = sort_related_symbols_by_stores(c2c_he, &num_rel);
+	if (!sorted)
+		return;
+
+	/* Create and populate child entries */
+	for (i = 0; i < num_rel; i++) {
+		struct hist_entry *child_he;
+		struct related_symbol *rel_sym = sorted[i];
+
+		/* Create child entry */
+		child_he = create_symbol_child_entry(he, rel_sym);
+		if (!child_he)
+			continue;
+
+		/* Populate grandchildren (cachelines) */
+		populate_cacheline_grandchildren(he, child_he, rel_sym);
+
+		/* Insert child into the parent's tree */
+		p = &root->rb_root.rb_node;
+		parent = NULL;
+		leftmost = true;
+		while (*p != NULL) {
+			parent = *p;
+			p = &parent->rb_right;
+			leftmost = false;
+		}
+
+		rb_link_node(&child_he->rb_node, parent, p);
+		rb_insert_color_cached(&child_he->rb_node, root, leftmost);
+
+		count++;
+	}
+
+	free(sorted);
+
+	/* Update parent's nr_rows and has_children flag */
+	if (count > 0) {
+		he->nr_rows = count;
+		he->has_children = true;
+	} else {
+		he->nr_rows = 0;
+		he->has_children = false;
+	}
+}
+
+/*
+ * ============================================================================
+ * Symbol Association Building
+ * ============================================================================
+ */
+
+/**
+ * build_symbol_associations - Build associations between symbols sharing cachelines
+ *
+ * Logic: When multiple symbols access the same cacheline (false sharing),
+ * they are considered related. The association count increases by 1 for
+ * each shared cacheline between two symbols.
+ */
+static void build_symbol_associations(void)
+{
+	struct rb_node *nd_sym;
+	struct hist_entry *he_sym;
+	struct c2c_hist_entry *c2c_he_sym;
+	int cl_idx;
+
+	/*
+	 * Algorithm:
+	 * 1. For each cacheline that has HITM events
+	 * 2. Look at its detailed access records
+	 * 3. Find all symbols that accessed it with HITM
+	 * 4. Create associations between these symbols
+	 */
+
+	/* Phase 1: Use cached index to find symbol conflicts efficiently */
+	/* Build cacheline index (will only build once due to internal flag) */
+	build_cacheline_symbol_index();
+
+	/* Iterate through cached index instead of rb-tree */
+	for (cl_idx = 0; cl_idx < c2c.cacheline_index_size; cl_idx++) {
+		struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[cl_idx];
+		struct c2c_hist_entry *c2c_he_cl = cl_entry->c2c_he_cl;
+		struct symbol_addr_pair {
+			struct symbol *sym;
+			uint64_t iaddr;
+		} *symbols_with_hitm = NULL;
+		int symbol_count = 0;
+		int symbol_capacity = 0;
+		struct symbol_access *sa;
+		int i, j;
+
+		/* Skip cachelines without HITM events */
+		if ((c2c_he_cl->stats.rmt_hitm + c2c_he_cl->stats.lcl_hitm) == 0)
+			continue;
+
+		/* Collect all (symbol, address) pairs that accessed this cacheline with HITM */
+		for (sa = cl_entry->symbol_accesses; sa; sa = sa->next) {
+			if (sa->sym && (sa->stats.rmt_hitm + sa->stats.lcl_hitm) > 0) {
+				uint64_t iaddr = sa->iaddr ? sa->iaddr : sa->sym->start;
+				/* Add (symbol, iaddr) pair to list if not already there */
+				bool found = false;
+
+				for (i = 0; i < symbol_count; i++) {
+					if (symbol_name_equal(symbols_with_hitm[i].sym, sa->sym) &&
+					    symbols_with_hitm[i].iaddr == iaddr) {
+						found = true;
+						break;
+					}
+				}
+
+				if (!found) {
+					if (symbol_count >= symbol_capacity) {
+						struct symbol_addr_pair *new_symbols;
+
+						symbol_capacity = symbol_capacity ? symbol_capacity * 2 : 4;
+						new_symbols = realloc(symbols_with_hitm,
+								    symbol_capacity * sizeof(struct symbol_addr_pair));
+						if (!new_symbols) {
+							/* Memory allocation failed, skip this symbol */
+							continue;
+						}
+						symbols_with_hitm = new_symbols;
+					}
+					if (symbols_with_hitm) {
+						symbols_with_hitm[symbol_count].sym = sa->sym;
+						symbols_with_hitm[symbol_count].iaddr = iaddr;
+						symbol_count++;
+					}
+				}
+			}
+		}
+
+		/* Create associations between all symbols that conflict on this cacheline */
+		if (symbol_count > 1) {
+			for (i = 0; i < symbol_count; i++) {
+				/* Find symbol entry in symbol_hists */
+				nd_sym = rb_first_cached(&c2c.symbol_hists.hists.entries);
+				while (nd_sym) {
+					he_sym = rb_entry(nd_sym, struct hist_entry, rb_node);
+					/* Match parent entry by BOTH symbol and code address */
+					{
+						uint64_t parent_iaddr = 0;
+
+						if (he_sym->mem_info)
+							parent_iaddr = mem_info__iaddr(he_sym->mem_info)->addr;
+						else if (he_sym->ms.sym)
+							parent_iaddr = he_sym->ms.sym->start;
+
+						if (symbol_name_equal(he_sym->ms.sym, symbols_with_hitm[i].sym) &&
+							parent_iaddr == symbols_with_hitm[i].iaddr) {
+							c2c_he_sym = container_of(he_sym, struct c2c_hist_entry, he);
+
+							/* Add all other symbols as related */
+							for (j = 0; j < symbol_count; j++) {
+								if (i != j) {
+									struct related_symbol *rel_sym;
+									bool exists = false;
+
+									/* Check if already added (compare both sym and iaddr) */
+									list_for_each_entry(rel_sym, &c2c_he_sym->related_symbols, list) {
+										if (symbol_name_equal(rel_sym->sym, symbols_with_hitm[j].sym) &&
+											rel_sym->iaddr == symbols_with_hitm[j].iaddr) {
+											exists = true;
+											break;
+										}
+									}
+
+									if (!exists) {
+										rel_sym = zalloc(sizeof(*rel_sym));
+										if (rel_sym) {
+											rel_sym->sym = symbols_with_hitm[j].sym;
+											rel_sym->iaddr = symbols_with_hitm[j].iaddr;
+											/* zalloc already zeros memory, no need for memset */
+											list_add_tail(&rel_sym->list, &c2c_he_sym->related_symbols);
+										}
+									}
+								}
+							}
+
+							/* Mark as having children */
+							if (!list_empty(&c2c_he_sym->related_symbols)) {
+								he_sym->has_children = true;
+								he_sym->unfolded = false;
+								he_sym->leaf = false;
+							}
+
+							break;
+						}
+					}
+					nd_sym = rb_next(nd_sym);
+				}
+			}
+		}
+
+		free(symbols_with_hitm);
+	}
+
+	/* Phase 2: Aggregate stats for related symbols using cached index */
+	nd_sym = rb_first_cached(&c2c.symbol_hists.hists.entries);
+	while (nd_sym) {
+		struct related_symbol *rel_sym;
+
+		he_sym = rb_entry(nd_sym, struct hist_entry, rb_node);
+		c2c_he_sym = container_of(he_sym, struct c2c_hist_entry, he);
+
+		/* For each related symbol, aggregate stats from shared cachelines using cached index */
+		list_for_each_entry(rel_sym, &c2c_he_sym->related_symbols, list) {
+			/* Precompute parent's iaddr once */
+			uint64_t parent_iaddr = he_sym->mem_info ?
+				mem_info__iaddr(he_sym->mem_info)->addr :
+				(he_sym->ms.sym ? he_sym->ms.sym->start : 0);
+
+			/* Use cached index instead of nested rb_next loops */
+			for (cl_idx = 0; cl_idx < c2c.cacheline_index_size; cl_idx++) {
+				struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[cl_idx];
+				struct symbol_access *sa_inner;
+				bool target_found = false;
+
+				/* First pass: check if target symbol accessed this cacheline */
+				for (sa_inner = cl_entry->symbol_accesses; sa_inner; sa_inner = sa_inner->next) {
+					if (symbol_name_equal(sa_inner->sym, he_sym->ms.sym) && sa_inner->iaddr == parent_iaddr) {
+						target_found = true;
+						break;
+					}
+				}
+
+				if (!target_found)
+					continue;
+
+				/* Second pass: aggregate related symbol stats */
+				for (sa_inner = cl_entry->symbol_accesses; sa_inner; sa_inner = sa_inner->next) {
+					if (symbol_name_equal(sa_inner->sym, rel_sym->sym) && sa_inner->iaddr == rel_sym->iaddr) {
+						c2c_add_stats(&rel_sym->stats, &sa_inner->stats);
+						c2c_add_cstats(&rel_sym->cstats, &sa_inner->cstats);
+					}
+				}
+			}
+		}
+
+		nd_sym = rb_next(nd_sym);
+	}
+
+	/* Phase 3: Create child entries for symbols with associations */
+	nd_sym = rb_first_cached(&c2c.symbol_hists.hists.entries);
+	while (nd_sym) {
+		he_sym = rb_entry(nd_sym, struct hist_entry, rb_node);
+		if (he_sym->has_children)
+			populate_symbol_children(he_sym);
+		nd_sym = rb_next(nd_sym);
+	}
+}
+
+/*
+ * ============================================================================
+ * Symbol Histogram Building
+ * ============================================================================
+ */
+
+/**
+ * calculate_symbol_total_cycles - Calculate total cycles for a single c2c_hist_entry
+ * @c2c_he: C2C histogram entry
+ *
+ * Returns the total cycles, using cached value if available
+ */
+static uint64_t calculate_symbol_total_cycles(struct c2c_hist_entry *c2c_he)
+{
+	uint64_t cycles_rmt, cycles_lcl, cycles_load, other_load;
+
+	/* Return cached value if available */
+	if (c2c_he->total_cycles_valid)
+		return c2c_he->total_cycles;
+
+	cycles_rmt = avg_stats(&c2c_he->cstats.rmt_hitm) * c2c_he->stats.rmt_hitm;
+	cycles_lcl = avg_stats(&c2c_he->cstats.lcl_hitm) * c2c_he->stats.lcl_hitm;
+	other_load = c2c_he->stats.load - c2c_he->stats.rmt_hitm - c2c_he->stats.lcl_hitm;
+	cycles_load = avg_stats(&c2c_he->cstats.load) * other_load;
+
+	/* Cache the result */
+	c2c_he->total_cycles = cycles_rmt + cycles_lcl + cycles_load;
+	c2c_he->total_cycles_valid = true;
+
+	return c2c_he->total_cycles;
+}
+
+/**
+ * get_total_cycles_all_symbols - Calculate total cycles for all symbols
+ *
+ * Returns the total cycles across all symbols, using cached value if available
+ */
+static uint64_t get_total_cycles_all_symbols(void)
+{
+	struct rb_node *nd;
+	uint64_t total_cycles = 0;
+
+	/* Use cached value if available to avoid O(n) scan per row */
+	if (c2c.symbol_total_cycles_valid)
+		return c2c.symbol_total_cycles;
+
+	nd = rb_first_cached(&c2c.symbol_hists.hists.entries);
+	while (nd) {
+		struct hist_entry *he = rb_entry(nd, struct hist_entry, rb_node);
+		struct c2c_hist_entry *c2c_he = container_of(he, struct c2c_hist_entry, he);
+
+		total_cycles += calculate_symbol_total_cycles(c2c_he);
+		nd = rb_next(nd);
+	}
+
+	c2c.symbol_total_cycles = total_cycles;
+	c2c.symbol_total_cycles_valid = true;
+	return total_cycles;
+}
+
+/**
+ * build_symbol_hists - Build symbol-level histograms from cacheline data
+ * @env: Perf environment containing symbol tables
+ *
+ * Returns: 0 on success, negative error code on failure
+ */
+int build_symbol_hists(struct perf_env *env)
+{
+	struct rb_node *next;
+	struct hist_entry *he_sym;
+	struct c2c_hist_entry *c2c_he_sym;
+	struct addr_location al;
+	struct perf_sample sample = {};
+	struct thread *synthetic_thread = NULL;
+	int ret, i;
+
+	/* Invalidate cached total cycles before (re)building symbol histograms */
+	c2c.symbol_total_cycles_valid = false;
+	c2c.symbol_total_cycles = 0;
+
+	next = rb_first_cached(&c2c.hists.hists.entries);
+
+	/* Initialize symbol hists with sort by iaddr (code address) and symbol */
+	ret = c2c_hists__init(&c2c.symbol_hists, "iaddr,symbol", 2, env);
+	if (ret)
+		return ret;
+
+	/* Setup output fields for symbol view - sorted by cycles percentage (descending) */
+	ret = c2c_hists__reinit(&c2c.symbol_hists,
+		"cycles_percent,total_stores,iaddr,symbol,cacheline_symbol",
+		"cycles_percent", env);
+	if (ret)
+		return ret;
+
+	/* Get first thread for consistent aggregation */
+	if (next) {
+		struct hist_entry *first_he = rb_entry(next, struct hist_entry, rb_node);
+
+		synthetic_thread = first_he->thread;
+	}
+
+	/* Build cacheline index (will only build once due to internal flag) */
+	build_cacheline_symbol_index();
+
+	if (c2c.cacheline_index_size == 0) {
+		ui__error("Cacheline index is empty, cannot build symbol hists\n");
+		return -1;
+	}
+
+	/* Directly create histogram entries from cached symbol_access data */
+	for (i = 0; i < c2c.cacheline_index_size; i++) {
+		struct cacheline_symbol_entry *cl_entry = &c2c.cacheline_index[i];
+		struct symbol_access *sa = cl_entry->symbol_accesses;
+
+		/* Process all symbol accesses for this cacheline */
+		while (sa) {
+			if (sa->sym) {
+				/* Create mem_info with proper instruction address for display */
+				struct mem_info *mi_display = mem_info__new();
+
+				if (mi_display) {
+					mem_info__iaddr(mi_display)->addr = sa->iaddr;
+					mem_info__iaddr(mi_display)->ms.maps = sa->maps;
+					mem_info__iaddr(mi_display)->ms.map = sa->map;
+					mem_info__iaddr(mi_display)->ms.sym = sa->sym;
+					/* Set data address to 0 for consistent display */
+					mem_info__daddr(mi_display)->addr = 0;
+				}
+
+				/* Create consistent address location for symbol aggregation */
+				addr_location__init(&al);
+				al.thread = synthetic_thread;
+				al.maps = sa->maps;
+				al.map = sa->map;
+				al.sym = sa->sym;
+				al.addr = sa->iaddr;
+				al.level = PERF_RECORD_MISC_KERNEL;
+				al.cpumode = PERF_RECORD_MISC_KERNEL;
+				al.cpu = 0;
+				al.socket = 0;
+				al.filtered = 0;
+
+				/* Create sample with consistent values */
+				sample.period = 1;
+				sample.weight = 1;
+				sample.ip = sa->iaddr;
+				sample.pid = synthetic_thread ? thread__pid(synthetic_thread) : 0;
+				sample.tid = synthetic_thread ? thread__tid(synthetic_thread) : 0;
+				sample.cpu = 0;
+				sample.time = 0;
+				sample.addr = 0;
+				sample.id = 0;
+
+				/* Add entry to histogram with mem_info for proper address display */
+				he_sym = hists__add_entry_ops(&c2c.symbol_hists.hists,
+							      &c2c_entry_ops,
+							      &al, NULL, NULL, mi_display,
+							      NULL, &sample, true);
+
+				addr_location__exit(&al);
+				if (mi_display)
+					mem_info__put(mi_display);
+
+				if (he_sym) {
+					c2c_he_sym = container_of(he_sym, struct c2c_hist_entry, he);
+
+					/* Copy aggregated stats from cached symbol_access */
+					c2c_he_sym->stats = sa->stats;
+					c2c_he_sym->cstats = sa->cstats;
+					c2c_add_stats(&c2c.symbol_hists.stats, &sa->stats);
+
+					hists__inc_nr_samples(&c2c.symbol_hists.hists, he_sym->filtered);
+					he_sym->hpp_list = &c2c.symbol_hists.list;
+				}
+			}
+			sa = sa->next;
+		}
+	}
+
+	/* Resort symbol hists */
+	hists__collapse_resort(&c2c.symbol_hists.hists, NULL);
+	hists__output_resort(&c2c.symbol_hists.hists, NULL);
+
+	/* Enable hierarchy support for symbol view to allow multi-level display */
+	c2c.symbol_hists.hists.symbol_filter_str = NULL;
+	c2c.symbol_hists.hists.socket_filter = -1;
+
+	/* Initialize the hist browser fields needed for hierarchy */
+	c2c.symbol_hists.hists.nr_hpp_node = 0;
+
+	/* Build symbol associations after hists are complete */
+	build_symbol_associations();
+
+	/* Precompute and cache total cycles to speed up percent rendering */
+	(void)get_total_cycles_all_symbols();
+
+	return 0;
+}
+
+/*
+ * ============================================================================
+ * Symbol Browser UI Components
+ * ============================================================================
+ */
 
 /**
  * c2c_symbol_browser__title - Generate title for symbol browser
@@ -145,4 +1179,3 @@ int c2c_symbol_browser__browse_cacheline_detail(struct c2c_symbol_browser *brows
 
 	return -1; /* Cacheline not found */
 }
-
